Move non-config concerns out of config/ into build-libs and src/lib
