+++ build-libs/rewrites.js
