+++ config/project-preview.json
@@ -1,4 +1,5 @@
 {
+  "extends": "base",
   "io_sites": {
     "max_static_paths": 9999,
     "revalidate": 360
