+++ src/components/navigation-header/index.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react'
 import Link from 'next/link'
-import { productSlugsToNames } from '../../../config/products'
+import { productSlugsToNames } from 'lib/products'
 import { ProductSlug } from 'types/products'
 import useCurrentPath from 'hooks/use-current-path'
 import { useCurrentProduct } from 'contexts'
