+++ src/lib/products.ts
